CLI registry list [--refresh] command and optional --refresh option for     the create commands

+++ dataset/GitHub_Java/quarkusio.quarkus/independent-projects/tools/registry-client/src/main/java/io/quarkus/registry/client/maven/MavenRegistryArtifactResolverWithCleanup.java
@@ -29,6 +29,13 @@ public class MavenRegistryArtifactResolverWithCleanup implements MavenRegistryAr
     }
 
     @Override
+    public Path findArtifactDirectory(Artifact artifact) throws BootstrapMavenException {
+        final LocalRepositoryManager localRepo = resolver.getSession().getLocalRepositoryManager();
+        return localRepo.getRepository().getBasedir().toPath().resolve(localRepo.getPathForLocalArtifact(artifact))
+                .getParent();
+    }
+
+    @Override
     public String getLatestVersionFromRange(Artifact artifact, String versionRange) throws BootstrapMavenException {
         return resolver.getLatestVersionFromRange(artifact, versionRange);
     }
@@ -48,7 +55,7 @@ public class MavenRegistryArtifactResolverWithCleanup implements MavenRegistryAr
      */
     protected static ArtifactResult resolveAndCleanupOldTimestampedVersions(MavenArtifactResolver resolver, Artifact artifact,
             boolean cleanupOldTimestampedVersions) throws BootstrapMavenException {
-        if (!cleanupOldTimestampedVersions) {
+        if (!artifact.isSnapshot() || !cleanupOldTimestampedVersions) {
             return resolver.resolve(artifact);
         }
 
@@ -63,7 +70,9 @@ public class MavenRegistryArtifactResolverWithCleanup implements MavenRegistryAr
         if (jsonDirContent != null && jsonDirContent.length > existingFiles.size()) {
             final String fileName = result.getArtifact().getFile().getName();
             for (File c : jsonDirContent) {
-                if (c.getName().length() > fileName.length() && c.getName().startsWith(artifact.getArtifactId())
+                if (c.getName().length() > fileName.length()
+                        && c.getName().startsWith(artifact.getArtifactId())
+                        && c.getName().endsWith(artifact.getClassifier())
                         && existingFiles.contains(c.getName())) {
                     c.deleteOnExit();
                 }
